backend/requirements.txt
click==8.0.1
enum-compat==0.0.3
eventlet==0.31.0
Flask==2.0.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-engineio==4.2.1
six==1.16.0
Werkzeug==2.0.1
line-bot-sdk==1.20.0
Flask-CORS==3.0.10
pymongo==3.12.0
